+++ test/setupTests.ts
@@ -1,5 +1,5 @@
 import { Utils } from "@uns/ark-crypto";
-import { NftMintDemandCertificationSigner } from "@uns/crypto";
+import { NftCertificationSigner } from "@uns/crypto";
 
 beforeEach(() => {
     // Reset all environment variables, which might be configured on this computer
@@ -16,7 +16,7 @@ afterEach(() => {
 beforeAll(async () => {
     // Handle ECDSA curve init ()
     try {
-        await new NftMintDemandCertificationSigner({ iss: "", sub: "", iat: 123, cost: Utils.BigNumber.ZERO }).sign(
+        await new NftCertificationSigner({ iss: "", sub: "", iat: 123, cost: Utils.BigNumber.ZERO }).sign(
             "path adult harsh access case slice source error protect brave reason okay",
         );
     } catch (_) {